chatgpt-telegram-elevenlabs-voice-assistant-main copy/requirements.txt
gTTS==2.3.1
moviepy==1.0.3
numpy==1.24.2
openai==0.27.0
python-telegram-bot==13.0
elevenlabslib==0.3.4
python-telegram-bot
torch
openai
moviepy
numpy
PIL
flask
flask_cors
autocast
diffusers
dotenv
playsound
requests
elevenlabslib